+++ doc/changelog.md
@@ -14,9 +14,11 @@ v0.2
 ===========
 - import node.js & nmp
 - import protobuf.js & socket.io
-- add simple PVP logic
 - use browerify & UglifyJS
 - add npm script to build & run
+- add simple PVP logic
+- add start UI
+- add leader board
 
 TODO
 =====